feat: implement the use of cropdb - ongoing
