+++ pyswap/core/db/cropdb.py
@@ -1,0 +1,140 @@
+"""
+# Crop database
+
+This module wraps the crop parameters database for WOFOST (A. de Wit). 
+"""
+
+from pathlib import Path
+from pprint import pprint
+from pyswap.libs import crop_params
+from pyswap.core.io.io_yaml import load_yaml
+from pydantic import BaseModel, computed_field
+import pandera as pa
+
+from typing import Literal
+
+
+class WOFOSTCropFile(BaseModel):
+    """Class for managing the content of a single WOFOST crop parameters file"""
+
+    yaml_content: dict
+
+    @computed_field(return_type=dict)
+    def metadata(self):
+        """Metadata of the yaml crop file"""
+        return self.yaml_content["Metadata"]
+
+    @computed_field(return_type=dict)
+    def ecotypes(self):
+        return list(self.yaml_content["CropParameters"]["EcoTypes"])
+
+    @computed_field(return_type=dict)
+    def genericc3(self):
+        """Get generic settings for C3 crop types - plants that bind CO2 into
+        3-phosphoglycerate having three carbon atoms. E.g., wheat, rice"""
+        return self.yaml_content["CropParameters"]["GenericC3"]
+
+    @computed_field(return_type=dict)
+    def genericc4(self):
+        """Get generic settings for C4 crop types - plants that bind CO2 into
+        oxaloacetate having four carbon atoms. E.g., maize, sugarcane"""
+        return self.yaml_content["CropParameters"]["GenericC4"]
+
+    @computed_field(return_type=dict)
+    def varieties(self):
+        return list(self.yaml_content["CropParameters"]["Varieties"])
+
+    def get_variety(self, variety: str, what: Literal["parameters", "metadata", "all"] = "parameters"):
+        variety = self.yaml_content["CropParameters"]["Varieties"][variety]
+        if what == "all":
+            return variety
+        elif what == "parameters":
+            return {k: v[0] for k, v in variety.items() if k != "Metadata"}
+        elif what == "metadata":
+            return variety["Metadata"]
+        else:
+            message = "Invalid value for 'what'. Choose from 'parameters', 'metadata', 'all'."
+            raise ValueError(message)
+
+    @staticmethod
+    def get_class_object(class_name: str):
+        import importlib
+        module = importlib.import_module("pyswap.components.crop")
+        if hasattr(module, class_name):
+            class_ = getattr(module, class_name)
+            return class_ if issubclass(class_, pa.DataFrameModel) else None
+        else:
+             None
+    
+    def format_tables(self, variety: dict):
+        for k, v in variety.items():
+            table_class = self.get_class_object(k)
+            if table_class:
+                cols = list(table_class.__annotations__.keys())
+                if isinstance(v, list):
+                    variety[k] = table_class.create(self._format_tables(v), columns=cols)
+        return {k: self._format_tables(v) if isinstance(v, list) else v for k, v in variety.items()}
+        # return variety
+    
+    @staticmethod
+    def _format_tables(table: list) -> dict:
+        """Format tables from YAML to a dictionary with two lists.
+        
+        In the YAML file, the tables seem to be formatted
+        in a way where the odd elements in the lists are one
+        column and the even elements are the other. This method
+        converts this format to a dictionary with two lists.
+        """
+        return {
+            "col1": table[::2],
+            "col2": table[1::2]
+        }
+
+
+def cropdb_to_pyswap(cropdb: WOFOSTCropFile):
+    """Convert a WOFOST crop database to pySWAP models
+    
+    The WOFOST crop database primarily contains information needed
+    for the following pySWAP models:
+    - CropSettings
+    - CO2Correction
+    - parts of Preparation
+    """
+    pass
+
+
+class WOFOSTCropDB(BaseModel):
+    """Simple class for managing crop parameters files.
+
+    Initially, it's meant to be used with A. de Wit's WOFOST crop parameters
+    database which is a collection of YAML files. However, it can be easily
+    extended to support other formats and databases when they emerge. All
+    methods should return the content as WOFOSTCropFile instances. This way
+    we ensure that whatever the source format is, the content is always
+    usable in the same way.
+
+    Attributes:
+        libdir: Path to the directory with crop parameters
+
+    Properties:
+        croptypes: List all available crop types (files in the directory)
+    
+    Methods:
+        load_crop_file: Load a specific crop file and return the content as a
+            WOFOSTCropFile instance
+    """
+
+    libdir: Path = crop_params
+
+    @computed_field(return_type=None)
+    def croptypes(self):
+        """Print the list of available files"""
+        pprint(load_yaml(crop_params / "crops.yaml")["available_crops"])
+        
+    def load_crop_file(self, crop: str):
+        """Load a specific crop file and return the content as a dictionary"""
+        path = self.libdir / f"{crop}" if crop.endswith(".yaml") else self.libdir / f"{crop}.yaml"
+        return WOFOSTCropFile(yaml_content=load_yaml(path))
+
+
+__all__ = ["WOFOSTCropDB"]
